remove getDynamicRouteParams and use parseObjectFromStringRecord, more parseServerSide*Params back to utils bc they were used in app too
