+++ packages/next-typesafe-url/src/pages.ts
@@ -5,12 +5,9 @@ import { z } from "zod";
 import { useState, useEffect } from "react";
 import {
   parseObjectFromParamString,
-  getDynamicRouteParams,
-  parseMapObject,
+  parseObjectFromStringRecord,
 } from "./utils";
 import type { UseParamsResult } from "./types";
-import type { ServerParseParamsResult } from "./types";
-import type { GetServerSidePropsContext } from "next";
 
 // ! Should ideally only be used in top level route component
 export function useRouteParams<T extends z.AnyZodObject>(
@@ -23,7 +20,7 @@ export function useRouteParams<T extends z.AnyZodObject>(
 
   useEffect(() => {
     if (router.isReady) {
-      const dynamicParams = getDynamicRouteParams(router.route, router.query);
+      const dynamicParams = parseObjectFromStringRecord(router.query);
       const validatedDynamicRouteParams = validator.safeParse(dynamicParams);
       if (validatedDynamicRouteParams.success) {
         setData(validatedDynamicRouteParams.data);
@@ -108,69 +105,5 @@ export function useSearchParams<T extends z.AnyZodObject>(
         error: undefined,
       };
     }
-  }
-}
-
-// takes gssp context.query
-export function parseServerSideSearchParams<T extends z.AnyZodObject>({
-  query,
-  validator,
-}: {
-  query: GetServerSidePropsContext["query"];
-  validator: T;
-}): ServerParseParamsResult<T> {
-  const parsedParams = parseMapObject(query);
-  const validatedDynamicSearchParams = validator.safeParse(parsedParams);
-  if (validatedDynamicSearchParams.success) {
-    return {
-      data: validatedDynamicSearchParams.data,
-      isError: false,
-      error: undefined,
-    };
-  } else {
-    return {
-      data: undefined,
-      isError: true,
-      error: validatedDynamicSearchParams.error,
-    };
-  }
-}
-
-// takes gssp context.params
-export function parseServerSideRouteParams<T extends z.AnyZodObject>({
-  params,
-  validator,
-}: {
-  params: GetServerSidePropsContext["params"];
-  validator: T;
-}): ServerParseParamsResult<T> {
-  if (!params) {
-    return {
-      data: undefined,
-      isError: true,
-      error: new z.ZodError([
-        {
-          path: [],
-          code: "custom",
-          message: "Params field of gSSP context is undefined",
-        },
-      ]),
-    };
-  }
-
-  const parsedParams = parseMapObject(params);
-  const validatedDynamicRouteParams = validator.safeParse(parsedParams);
-  if (validatedDynamicRouteParams.success) {
-    return {
-      data: validatedDynamicRouteParams.data,
-      isError: false,
-      error: undefined,
-    };
-  } else {
-    return {
-      data: undefined,
-      isError: true,
-      error: validatedDynamicRouteParams.error,
-    };
   }
 }
